Time: 3 ms (17.72%) | Memory: 8 MB (38.88%) - LeetSync

+++ 374-guess-number-higher-or-lower/guess-number-higher-or-lower.cpp
@@ -10,7 +10,7 @@
 class Solution {
 public:
     int guessNumber(int n) {
-        long long low=1,high=n;
+        long low=1,high=n;
         long long mid=(low+high)/2;
 
         while(low<=high)
